Give `GFp_*` modular arithmetic primitives more general names.
These aren't just useful for finite fields over primes GFp. They can
be used for arithmetic in the ring Z/Zn where n isn't prime, as is
the case for RSA arithmetic modulo the public modulus `n`.

The prefix `RING_` is now used for the exported symbols in this module.
Now it is clear now that we're not going to need to disambiguate the
types in this module with other libraries' types, so remove the `GFp_`
prefixes.